requirements.txt
flask==3.1.1
flask-limiter==3.8.0  # For rate limiting
requests==2.32.4
redis==6.3.0
rq==2.4.1
pytest==8.4.1
six==1.17.0
pytest-flask==1.3.0  # For Flask test client
pytest-mock==3.12.0  # For mocking in tests
fakeredis==2.30.3    # For mocking Redis in tests
google-genai==1.28.0
python-dotenv==1.1.1
